pyproject.toml
[tool.poetry]
name = "pic-analyzer"
version = "0.1.0"
description = ""
authors = ["Guillaume <7831777+GuillaumeOj@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "^10.0.1"
fastapi = {extras = ["all"], version = "^0.104.0"}
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-multipart = "^0.0.6"

[tool.poetry.dev-dependencies]
black = "^23.9.1"
flake8 = "^6.1.0"
isort = "^5.12.0"
docformatter = "^1.7.5"
ipdb = "^0.13.13"
ipython = "^8.16.1"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
